put cudnn_lstm class in models.py
